+++ docs/tailwind.config.ts
@@ -6,7 +6,7 @@ import { type Config } from 'tailwindcss';
 export default {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     presets: [require('@frontify/fondue-tokens/tailwind')],
-    content: ['./src/**/*.{js,ts,tsx}', './node_modules/@frontify/fondue-components/src/**/*.{js,ts,tsx}'],
+    content: ['./src/**/*.{js,ts,tsx}'],
     plugins: [
         tailwindForms({
             strategy: 'class',